Move network annotation processor tests.
