+++ modules/network/src/test/java/org/apache/ignite/internal/network/processor/InheritedMessage.java
